types

+++ genetic_decomposition/area_graph.h
@@ -15,12 +15,12 @@ public:
     /// <summary>
     /// 
     /// </summary>
-    vector<vector<int>> inc;
+    vector<vector<size_t>> inc;
 
     /// <summary>
     /// 
     /// </summary>
-    vector<vector<int>> es;
+    vector<vector<size_t>> es;
 
     /// <summary>
     /// 
@@ -53,8 +53,8 @@ public:
     /// <param name="a"></param>
     /// <param name="b"></param>
     void
-    link_two_nodes(int a,
-                   int b);
+    link_two_nodes(size_t a,
+                   size_t b);
 
     /// <summary>
     /// 
@@ -62,8 +62,8 @@ public:
     /// <param name="rows"></param>
     /// <param name="cols"></param>
     void
-    init_from_rectangular_area(int rows,
-                               int cols);
+    init_from_rectangular_area(size_t rows,
+                               size_t cols);
 
     /// <summary>
     ///
